// asb-good-foods-user/src/app/common/models/profile.model.ts
export class Profile {
    email?: string;
    mobile?:number;
    date?: any;
    userDocId?:string;
    orderItems?:any;
    fullName: string;
    addressLine2: string;
    pincode: number;
    addressLine1: string;
    city: string;
    state: string;
    sameAddress: boolean;
    sfullName: string;
    saddressLine2: string;
    smobile: number;
    spinCode: number;
    saddressLine1: string;
    scity: string;
    sstate: string;
    profileUrl: string;
    addresslabel: string;
    userType : string;
    userCompanyName: string;
    companyGst : string;
    companyCin : string;
    companyPan : string;
    businessPanUrl: string;
    supportingDocUrl:string;
    // productName?: string;
    // description?: string;
    // productPrice?: number;
    // imageUrl?: string;
    // subtotal?: number;
    // quantity? : number;
    // standard?: boolean;
    // tracked?: boolean;
    // userId?: string;
    // orderStatus?: string;
    // orderId? : number;
    // gsttotal?: number;
    // gst?: number;
    // total: number;
    // shipping: number;
}